Restructure project's database package
