--- ExamApp/src/main/java/com/OBJ2100/ExamApp/db/dao/implementations/OfficeDaoImpl.java
@@ -1,5 +1,0 @@
-package com.OBJ2100.ExamApp.db.dao.implementations;
-
-public class OfficeDaoImpl {
-
-}
