+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/dao/jdbc/JdbcOrderDao.java
@@ -1,0 +1,5 @@
+package com.OBJ2100.ExamApp.db.dao.jdbc;
+
+public class JdbcOrderDao {
+
+}
