+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/db/dao/jdbc/JdbcProductLineDao.java
@@ -1,0 +1,5 @@
+package com.OBJ2100.ExamApp.db.dao.jdbc;
+
+public class JdbcProductLineDao {
+
+}
