--- ExamApp/src/main/java/com/OBJ2100/ExamApp/db/entities/OrderDetails.java
@@ -1,5 +1,0 @@
-package com.OBJ2100.ExamApp.db.entities;
-
-public class OrderDetails {
-
-}
