+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/entities/Order.java
@@ -1,4 +1,4 @@
-package com.OBJ2100.ExamApp.db.entities;
+package com.OBJ2100.ExamApp.entities;
 
 import java.util.Date;
