+++ ExamApp/src/main/java/com/OBJ2100/ExamApp/entities/OrderDetails.java
@@ -1,0 +1,5 @@
+package com.OBJ2100.ExamApp.entities;
+
+public class OrderDetails {
+
+}
